services/server_node/requirements.txt
flask
pycryptodome
requests
psycopg2-binary